+++ ephFrame/view/element/errorMessage.php
@@ -1,3 +1,0 @@
-<p class="error">
-	<?php echo @$message ?>
-</p>