+++ ephFrame/view/error/missingController.php
@@ -3,9 +3,9 @@
 <h1>Missing Controller <q><?php echo $controllerName ?></q></h1>
 <p class="error">
 	Please create the controller class in
-	<q><?php echo APP_LIB_DIR.'controller/'.$controllerName.'.php' ?></q>
+	<code><?php echo APP_LIB_DIR.'controller/'.$controllerName.'.php' ?></code>
 </p>
 <p class="hint">
 	You can edit this error message by creating your own view for it in
-	<q><?php echo VIEW_DIR ?>error/<?php echo basename(__FILE__) ?></q>
+	<code><?php echo VIEW_DIR.'error/'.basename(__FILE__) ?></code>
 </p>